# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: VerdictAI REST API
  description: |
    # VerdictAI: Frictionless Dispute & Chargeback Resolution
    FastAPI REST Gateway exposing endpoints for Card Members (Mobile), Merchants (Web), and Dispute-Ops Analysts/Admins (Web).
    
    Developed by **Darshan Prajapati** (`202512026`) — Backend Engineer (Reasoning & APIs).
    
    ### Key Features:
    - **Dispute Lifecycle Management:** Submit, track, transition, and seal dispute files with ACID state guarantees.
    - **Polymorphic Evidence Storage:** Attach receipts, tracking data, communication logs with automated SHA-256 checksums.
    - **Explainable AI (XAI) Reasoning:** Plain-language rationales and factor breakdowns powered by Google Gemini.
    - **Human-in-the-Loop Admin Overrides:** Filterable review queue and audited overrides with mandatory reasoning.
  version: 1.0.0
  contact:
    name: Darshan Prajapati
    email: 202512026@daiict.ac.in
servers:
  - url: http://localhost:8000
    description: Local Development Server
  - url: https://api.verdictai.render.com
    description: Production Cloud Gateway

paths:
  /health:
    get:
      tags:
        - System
      summary: Health Check
      description: Returns service health status and timestamp.
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  service:
                    type: string
                    example: "VerdictAI API Gateway"
                  timestamp:
                    type: string
                    format: date-time

  /api/v1/system/status:
    get:
      tags:
        - System
      summary: System & Component Status
      description: Checks database connectivity, ML model availability, and reasoning engine readiness.
      responses:
        '200':
          description: Detailed system diagnostics.

  /api/v1/transactions/cardholder/{cardholder_id}:
    get:
      tags:
        - Transactions
      summary: Get Eligible Transactions for Card Member
      description: Returns recent card transactions available for dispute filing.
      parameters:
        - name: cardholder_id
          in: path
          required: true
          schema:
            type: string
          example: "usr_001"
      responses:
        '200':
          description: List of transactions.

  /api/v1/transactions/{transaction_id}:
    get:
      tags:
        - Transactions
      summary: Get Transaction Details
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction detail.
        '404':
          description: Transaction not found.

  /api/v1/disputes:
    post:
      tags:
        - Disputes
      summary: File a New Dispute (Card Member)
      description: Creates a new dispute case, registers transaction metadata, calculates SLA deadline, and initializes audit trail.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisputeCreateRequest'
      responses:
        '201':
          description: Dispute successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedCaseFile'
        '400':
          description: Invalid dispute payload or validation error.
        '404':
          description: Referenced transaction not found.

    get:
      tags:
        - Disputes
      summary: List / Search Disputes
      description: Filterable and paginated list of disputes for card members, merchants, and operations analysts.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DisputeStatus'
        - name: cardholder_id
          in: query
          required: false
          schema:
            type: string
        - name: merchant_id
          in: query
          required: false
          schema:
            type: string
        - name: reason
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DisputeReason'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated disputes response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeListResponse'

  /api/v1/disputes/{dispute_id}:
    get:
      tags:
        - Disputes
      summary: Get Unified Case File
      description: Retrieves complete case document aggregating PostgreSQL metadata, MongoDB evidence payloads, and audit history.
      parameters:
        - name: dispute_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unified Case File.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedCaseFile'
        '404':
          description: Dispute not found.

  /api/v1/disputes/{dispute_id}/status:
    patch:
      tags:
        - Disputes
      summary: Transition Dispute Lifecycle State
      description: Validates and executes state transitions via the Dispute State Machine with audit logging.
      parameters:
        - name: dispute_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateTransitionRequest'
      responses:
        '200':
          description: State successfully updated.
        '400':
          description: Invalid state transition requested.
        '404':
          description: Dispute not found.

  /api/v1/disputes/{dispute_id}/audit-trail:
    get:
      tags:
        - Disputes
      summary: Get Cryptographic Audit Trail
      description: Returns the tamper-evident SHA-256 hash-chained audit log for the case.
      parameters:
        - name: dispute_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ordered list of audit entries and chain verification status.

  /api/v1/evidence/upload:
    post:
      tags:
        - Evidence
      summary: Upload Evidence File (PDF, JPG, PNG)
      description: Ingests an evidence file, computes SHA-256 checksum, attaches to MongoDB case file, and advances dispute state.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - dispute_id
                - evidence_type
                - source
                - file
              properties:
                dispute_id:
                  type: string
                evidence_type:
                  $ref: '#/components/schemas/EvidenceType'
                source:
                  $ref: '#/components/schemas/EvidenceSource'
                actor:
                  type: string
                  default: "SYSTEM_UPLOAD"
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Evidence ingested and verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidencePayloadModel'
        '400':
          description: File size exceeds threshold (>10MB for cardholders, >25MB for merchants) or unsupported MIME type.

  /api/v1/evidence/payload:
    post:
      tags:
        - Evidence
      summary: Attach Structured Evidence Payload
      description: Ingests structured JSON evidence (e.g. carrier GPS data, invoice details, dialogue logs).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StructuredEvidenceRequest'
      responses:
        '201':
          description: Evidence attached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidencePayloadModel'

  /api/v1/evidence/statement:
    post:
      tags:
        - Evidence
      summary: Submit Cardholder or Merchant Statement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatementSubmissionRequest'
      responses:
        '200':
          description: Statement recorded.

  /api/v1/disputes/{dispute_id}/evidence:
    get:
      tags:
        - Evidence
      summary: List Evidence for a Dispute
      parameters:
        - name: dispute_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of evidence items.

  /api/v1/admin/queue:
    get:
      tags:
        - Admin & Operations
      summary: Get Prioritized Admin Review Queue
      description: Returns cases awaiting manual review, sorted by confidence score (< 50%) and SLA deadline.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            default: "MANUAL_REVIEW_QUEUE"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Paginated queue of cases.

  /api/v1/admin/disputes/{dispute_id}/override:
    post:
      tags:
        - Admin & Operations
      summary: Override Automated Dispute Decision
      description: Dispute-Ops Admin overrides an automated decision. Requires mandatory written reason for audit logging (FR-27, AC-13).
      parameters:
        - name: dispute_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOverrideRequest'
      responses:
        '200':
          description: Decision overridden and logged.
        '400':
          description: Missing mandatory written reason or invalid decision.
        '404':
          description: Dispute not found.

  /api/v1/admin/disputes/{dispute_id}/assign:
    post:
      tags:
        - Admin & Operations
      summary: Assign Case to Analyst
      parameters:
        - name: dispute_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - analyst_id
                - analyst_name
              properties:
                analyst_id:
                  type: string
                analyst_name:
                  type: string
      responses:
        '200':
          description: Case assigned.

components:
  schemas:
    DisputeStatus:
      type: string
      enum:
        - SUBMITTED
        - EVIDENCE_PENDING
        - EVIDENCE_INGESTED
        - IN_ANALYSIS
        - SCORING_EVALUATED
        - AUTO_RESOLVED
        - MANUAL_REVIEW_QUEUE
        - ADMIN_OVERRIDDEN
        - RESOLUTION_NOTIFIED
        - CLOSED
        - REJECTED

    DisputeReason:
      type: string
      enum:
        - FRAUD_UNRECOGNIZED_CHARGE
        - PRODUCT_NOT_RECEIVED
        - PRODUCT_DAMAGED_OR_DEFECTIVE
        - SUBSCRIPTION_CANCELLED_CHARGED
        - DUPLICATE_PROCESSING
        - INCORRECT_AMOUNT_CHARGED

    EvidenceSource:
      type: string
      enum:
        - CARDHOLDER
        - MERCHANT
        - PAYMENT_GATEWAY
        - COURIER_CARRIER
        - SYSTEM_AUTOMATION

    EvidenceType:
      type: string
      enum:
        - RECEIPT_INVOICE
        - COURIER_TRACKING
        - COMMUNICATION_LOG
        - REFUND_POLICY_TERMS
        - BANK_STATEMENT
        - IDENTITY_VERIFICATION
        - SYSTEM_AUDIT_PROOF

    ResolutionOutcome:
      type: string
      enum:
        - FAVOR_CARDHOLDER
        - FAVOR_MERCHANT
        - SPLIT_LIABILITY
        - MERCHANT_ACCEPTED

    DisputeCreateRequest:
      type: object
      required:
        - transaction_id
        - cardholder_id
        - dispute_reason
        - disputed_amount
      properties:
        transaction_id:
          type: string
          example: "txn_apple_1099"
        cardholder_id:
          type: string
          example: "usr_alice_01"
        dispute_reason:
          $ref: '#/components/schemas/DisputeReason'
        disputed_amount:
          type: number
          format: float
          example: 1249.50
        cardholder_statement:
          type: string
          example: "The package was never delivered to my residence. Tracking says delivered but no parcel found."
        merchant_sla_hours:
          type: integer
          default: 48

    StateTransitionRequest:
      type: object
      required:
        - target_status
        - actor
        - reason
      properties:
        target_status:
          $ref: '#/components/schemas/DisputeStatus'
        actor:
          type: string
          example: "DISPUTE_OPS_ANALYST:alex"
        reason:
          type: string
          example: "Evidence review completed, moving to analysis"

    StructuredEvidenceRequest:
      type: object
      required:
        - dispute_id
        - evidence_type
        - source
        - raw_payload
      properties:
        dispute_id:
          type: string
        evidence_type:
          $ref: '#/components/schemas/EvidenceType'
        source:
          $ref: '#/components/schemas/EvidenceSource'
        actor:
          type: string
          default: "SYSTEM"
        file_name:
          type: string
        raw_payload:
          type: object
          additionalProperties: true

    StatementSubmissionRequest:
      type: object
      required:
        - dispute_id
        - statement_type
        - statement_text
        - author
      properties:
        dispute_id:
          type: string
        statement_type:
          type: string
          enum: [CARDHOLDER, MERCHANT]
        statement_text:
          type: string
        author:
          type: string

    AdminOverrideRequest:
      type: object
      required:
        - override_decision
        - mandatory_reason
        - admin_id
      properties:
        override_decision:
          $ref: '#/components/schemas/ResolutionOutcome'
        mandatory_reason:
          type: string
          minLength: 10
          example: "Carrier geofence GPS discrepancy confirmed parcel dropped 0.4 miles away from customer address."
        admin_id:
          type: string
          example: "adm_marcus_01"
        admin_name:
          type: string
          example: "Marcus Chen"

    TransactionSummary:
      type: object
      properties:
        transaction_id:
          type: string
        amount:
          type: number
        currency:
          type: string
        merchant_id:
          type: string
        merchant_name:
          type: string
        cardholder_id:
          type: string
        cardholder_name:
          type: string
        payment_method:
          type: string
        transaction_timestamp:
          type: string
          format: date-time

    CaseFileHeader:
      type: object
      properties:
        case_file_id:
          type: string
        dispute_id:
          type: string
        case_reference_number:
          type: string
        current_status:
          $ref: '#/components/schemas/DisputeStatus'
        dispute_reason:
          $ref: '#/components/schemas/DisputeReason'
        disputed_amount:
          type: number
        currency:
          type: string
        created_at:
          type: string
          format: date-time
        sla_deadline:
          type: string
          format: date-time
        is_sealed:
          type: boolean
        sealed_at:
          type: string
          format: date-time

    EvidencePayloadModel:
      type: object
      properties:
        evidence_id:
          type: string
        dispute_id:
          type: string
        evidence_type:
          $ref: '#/components/schemas/EvidenceType'
        source:
          $ref: '#/components/schemas/EvidenceSource'
        file_name:
          type: string
        mime_type:
          type: string
        file_size_bytes:
          type: integer
        sha256_checksum:
          type: string
        uploaded_at:
          type: string
          format: date-time
        raw_payload:
          type: object
        nlp_extracted_entities:
          type: object
        ocr_extracted_text:
          type: string
        confidence_rating:
          type: number

    UnifiedCaseFile:
      type: object
      properties:
        header:
          $ref: '#/components/schemas/CaseFileHeader'
        transaction:
          $ref: '#/components/schemas/TransactionSummary'
        cardholder_statement:
          type: string
        merchant_response_statement:
          type: string
        evidence_items:
          type: array
          items:
            $ref: '#/components/schemas/EvidencePayloadModel'
        case_hash_sha256:
          type: string
        audit_chain_length:
          type: integer
        resolution:
          type: object

    DisputeListResponse:
      type: object
      properties:
        total_count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/UnifiedCaseFile'